+++ ipatests/pytest_ipa/integration/config.py
@@ -92,11 +92,11 @@ class Config(pytest_multihost.config.Config):
 
     @classmethod
     def from_env(cls, env):
-        from ipatests.pytest_plugins.integration.env_config import config_from_env
+        from ipatests.pytest_ipa.integration.env_config import config_from_env
         return config_from_env(env)
 
     def to_env(self, **kwargs):
-        from ipatests.pytest_plugins.integration.env_config import config_to_env
+        from ipatests.pytest_ipa.integration.env_config import config_to_env
         return config_to_env(self, **kwargs)
 
 
@@ -124,7 +124,7 @@ class Domain(pytest_multihost.config.Domain):
             raise LookupError(self.type)
 
     def get_host_class(self, host_dict):
-        from ipatests.pytest_plugins.integration.host import Host, WinHost
+        from ipatests.pytest_ipa.integration.host import Host, WinHost
 
         if self.type == 'IPA':
             return Host
@@ -159,9 +159,9 @@ class Domain(pytest_multihost.config.Domain):
 
     @classmethod
     def from_env(cls, env, config, index, domain_type):
-        from ipatests.pytest_plugins.integration.env_config import domain_from_env
+        from ipatests.pytest_ipa.integration.env_config import domain_from_env
         return domain_from_env(env, config, index, domain_type)
 
     def to_env(self, **kwargs):
-        from ipatests.pytest_plugins.integration.env_config import domain_to_env
+        from ipatests.pytest_ipa.integration.env_config import domain_to_env
         return domain_to_env(self, **kwargs)
