+++ ipatests/test_integration/test_commands.py
@@ -13,7 +13,7 @@ import textwrap
 from ipaplatform.paths import paths
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 
 
 class TestIPACommand(IntegrationTest):
