+++ ipatests/test_integration/test_dnssec.py
@@ -13,7 +13,7 @@ import dns.name
 import time
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 from ipaplatform.paths import paths
 
 logger = logging.getLogger(__name__)
