+++ ipatests/test_integration/test_forced_client_reenrollment.py
@@ -26,7 +26,7 @@ from ipaplatform.paths import paths
 import pytest
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 
 logger = logging.getLogger(__name__)
