+++ ipatests/test_integration/test_http_kdc_proxy.py
@@ -5,7 +5,7 @@
 from __future__ import absolute_import
 
 import six
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 from ipatests.test_integration.base import IntegrationTest
 from ipaplatform.paths import paths
