+++ ipatests/test_integration/test_service_permissions.py
@@ -20,7 +20,7 @@
 import os
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 
 
 class TestServicePermissions(IntegrationTest):
