+++ ipatests/test_integration/test_trust.py
@@ -23,7 +23,7 @@ import nose
 import re
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 from ipaplatform.paths import paths
