+++ ipatests/test_integration/test_upgrade.py
@@ -7,7 +7,7 @@ Module provides tests to verify that the upgrade script works.
 """
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 
 
 class TestUpgrade(IntegrationTest):
