+++ ipatests/test_integration/test_user_permissions.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import
 
 from ipaplatform.paths import paths
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_ipa.integration import tasks
 
 
 class TestUserPermissions(IntegrationTest):
